More cleaning up. Also combined all protocols into one hex.
At power up channel 8 is scanned: 

If it is LOW (1000us +10us) normal Flysky modulation is used and the RED LED is lit.

If channel 8 is NEUTRAL (1500us +-10us) Flysky modulation with quadcopter mods is used and the YELLOW LED is lit. With these mods enabled channels 5,6,and 7 take on extra rolls, LEDs, flips, camera, etc.

If channel 8 is high (2000us -10us) Hubsan modulation is used and the GREEN LED is lit.
Changing channel 8 after power up does not alter the modulation.

THIS IS ALL UNTESTED ! I'll debug it later today, hopefully it won't need it.